Coalesce contiguous sequences of text nodes when parsing SXML.
Some SAX parsers emit long runs of 'characters' events, often with very
short payloads, particularly when there are large numbers of entity
references in the source document. This change causes the resulting runs
of separate text nodes to be coalesced into a single string.

This is done "lazily", that is, the runs are allowed to accumulate on
the stack until another type of event appears, at which point the run is
concatenated all at one time.