Support for compiled-sxml macro.
This macro pre-compiles a literal SXML form that will be used as input
to JAXP. Non-constant expressions are permitted as long as they do not
modify the structure of the resulting document, i.e. they are always
compiled as text nodes. Attributes must be a literal map for now, but
may contain expressions inside them.