Convert read-sxml input adaptation into a protocol.
Define a protocol XMLReadable with two methods, make-xmlreader and
xmlreader-parse-arg, which replaces two multimethods. Extend various
builtins to XMLReadable. This means other kinds of things can be passed
to sax/read-sxml by making them extend XMLReadable.

The two methods return an instance of XMLReader which will read the
input, and the instance of InputSource that will be passed to that
XMLReader's parse() method.