Convert the transform result adaptation to protocols.
The TransformResult protocol allows objects to produce a Result object
along with an arbitrary piece of state which extends TransformResult.

The state is used to produce the result "value", either by simply
handing back a mutated Java object, unwrapping it, or by extracting a
value generated by the API mutating the Result.