Apply the proper xmlns declaration in transform.xslt/stylesheet.
This fixes a bug where the stylesheet gets filtered into an event-seq
before being passed to transform/compile-template--the default binding
does not take effect and the stylesheet fails to compile.

It doesn't work because default declarations are applied in
sax/sax-event-seq, which isn't called for things that are already an
event seq.